Change proximity behaviour for devices w/ fixed keyboard
Imho the actual behaviour of Phone (force the screen to stay On when the Keyboard is open) is wrong, at least in the case of the HTC Chacha (with a fixed open keyboard). Maybe it has sense only for devices with sliding keyboard.

This patch to PhoneApp.java is to behave diferently if the device has a fixed / always open physical keyboard (by reading ro.product.has_fixed_keyboard value), and don't care about if keyboard is open/hidden in the case.

Change-Id: Ia3b635dca889508404d21a29c014cc5b0413b8b0